requirements.txt
streamlit
google-genai
python-dotenv
PyMuPDF
requests
beautifulsoup4
reportlab
pandas
